requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
